OrigenX/BackEnd/requirements.txt
# Marketplace Cafetero – Backend Python (FastAPI)
# Instalar con: pip install -r requirements.txt

# Framework HTTP
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Firebase Admin SDK (Auth, Firestore, Storage)
firebase-admin==6.5.0

# Validación de datos
pydantic==2.7.4
pydantic[email]==2.7.4

# Variables de entorno
python-dotenv==1.0.1

# Utilidades HTTP
httpx==0.27.0

# Soporte para multipart/form-data (requerido por FastAPI para UploadFile)
python-multipart==0.0.9

# Seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# CORS
# (incluido en fastapi/starlette)

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
hypothesis==6.104.2
pytest-cov==7.1.0
